scripts/requirements.txt
gspread==5.12.4
oauth2client==4.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
reportlab==4.0.7
schedule==1.2.0
twilio==8.10.3
requests==2.31.0